requirements.txt
alabaster==0.7.12
astroid==2.11.5
Babel==2.10.1
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
crcmod==1.7
dill==0.3.4
docutils==0.17.1
idna==3.3
imagesize==1.3.0
isort==5.10.1
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
numpy==1.22.3
packaging==21.3
platformdirs==2.5.2
Pygments==2.12.0
pylint==2.13.9
pyparsing==3.0.9
pytz==2022.1
pyusb==1.2.1
requests==2.27.1
snowballstemmer==2.2.0
Sphinx==4.5.0
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tomli==2.0.1
urllib3==1.26.9
wrapt==1.14.1
